Replace `unsigned int` with `size_t`

+++ 00_Libft/include/ft_string.h
@@ -6,7 +6,7 @@
 /*   By: kijsong <kijsong@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/23 09:29:49 by kijsong           #+#    #+#             */
-/*   Updated: 2022/09/09 01:53:22 by kijsong          ###   ########.fr       */
+/*   Updated: 2022/09/09 02:09:08 by kijsong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,11 +33,11 @@ int		ft_strcmp(const char *s1, const char *s2);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 
 char	*ft_strdup(const char *s1);
-char	*ft_substr(const char *s, unsigned int start, size_t len);
+char	*ft_substr(const char *s, size_t start, size_t len);
 char	*ft_strjoin(const char *s1, const char *s2);
 char	*ft_strtrim(const char *s1, const char *set);
 char	**ft_split(const char *s, char c);
-char	*ft_strmapi(const char *s, char (*f)(unsigned int, char));
-void	ft_striteri(char *s, void (*f)(unsigned int, char *));
+char	*ft_strmapi(const char *s, char (*f)(size_t, char));
+void	ft_striteri(char *s, void (*f)(size_t, char *));
 
 #endif
